Revert "카테고리 완료"
