--- app/landing/animations/fadeInAnimation.ts
@@ -1,18 +1,0 @@
-import { Variants } from 'framer-motion';
-
-const fadeInVariants: Variants = {
-  hidden: {
-    opacity: 0,
-    translateY: 0,
-  },
-  visible: {
-    width: '100%',
-    opacity: 1,
-    translateY: -100,
-    transition: {
-      duration: 1,
-    },
-  },
-};
-
-export default fadeInVariants;
