+++ app/map/_components/Autocomplete/AutocompleteMini.module.scss
@@ -1,9 +1,9 @@
-@import 'styles/global-styles/index';
+@import "styles/global-styles/index";
 
 //변수적용이 안되는중 ㅠ 억까하지마라 진짜
 .container {
-  width: 30rem;
-  max-width: 30rem;
+  width: 60rem;
+  max-width: 60rem;
 
   .input-container {
     display: flex;
@@ -15,18 +15,19 @@
     gap: 1.2rem;
     flex-shrink: 0;
 
-    border-bottom: 1.6px solid var(--BDS-Grayscale-6, #bfbfbf);
-    background: var(--BDS-Grayscale-1, #fff);
+    border-radius: 1rem 3.2rem 3.2rem 1rem;
+    border: 2.4px solid var(--BDS-Primary-5, #FF9456);
+    background: var(--BDS-Grayscale-1, #FFF);
+
 
     input {
       border: none;
       width: 100%;
 
-      &:focus {
-        border: none;
+      caret-color: var(--BDS-Primary-5, #FF9456);
 
-        border-bottom: 2.4px solid var(--BDS-Primary-3, #ffc7a6);
-        background: var(--BDS-Grayscale-1, #fff);
+      &:focus {
+        outline: 0;
       }
     }
 
@@ -35,13 +36,13 @@
       height: 1.8rem;
       flex-shrink: 0;
       stroke-width: 2.4px;
-      stroke: var(--BDS-Primary-5, #ff9456);
+      stroke: var(--BDS-Primary-5, #FF9456);
     }
   }
 
   ul {
-    // z-index: $above + $z-index-root;
-    // position: absolute;
+    z-index: $above +$z-index-root;
+    position: absolute;
     background-color: whitesmoke;
     width: inherit;
     cursor: pointer;
@@ -54,8 +55,8 @@
       gap: 1rem;
 
       &:hover {
-        border-bottom: 1px solid var(--BDS-Grayscale-5, #d9d9d9);
-        background: var(--BDS-Primary-1, linear-gradient(90deg, #fff2ea 0%, #fff0f8 100%));
+        border-bottom: 1px solid var(--BDS-Grayscale-5, #D9D9D9);
+        background: var(--BDS-Primary-1, linear-gradient(90deg, #FFF2EA 0%, #FFF0F8 100%));
       }
     }
   }
