+++ components/GlobalNavigationBar/NotificationModal/NotificationItem.tsx
@@ -15,6 +15,8 @@ export default function NotificationItem({ notification }: { notification: Notif
 
   const handleClickDeleteNotification = () => {
     deleteNotificationMutate(notification.notificationId);
+
+    console.log(notification.notificationId, '번 알림을 삭제합니다.');
   };
 
   return (
