+++ components/Modal/WritePost/PostModalInput.module.scss
@@ -4,7 +4,6 @@
   @include flex-column;
   align-items: flex-start;
   gap: 0.4rem;
-  width: 30rem;
 }
 
 .message {
