+++ package.json
@@ -39,7 +39,6 @@
     "axios": "^1.6.7",
     "classnames": "^2.5.1",
     "d3-interpolate": "^3.0.1",
-    "framer-motion": "^11.2.0",
     "moment": "^2.30.1",
     "next": "14.2.3",
     "react": "^18",
