--- public/fonts/bagel-fat-one/Bagel_Fat_One.css
@@ -1,6 +1,0 @@
-@font-face {
-  font-family: 'Bagel Fat One';
-  font-weight: 400;
-  font-style: normal;
-  src: url(./BagelFatOne-Regular.ttf) format('opentype');
-}
