+++ queries/useUserQueries.ts
@@ -97,6 +97,7 @@ export function useDeleteUser() {
 
       logout();
       deleteCookies();
+      localStorage.clear();
     },
     onError: (error) => {
       console.error('회원탈퇴 실패ㅜㅜ:', error);
